requirements.txt
click==8.1.7
duplicity==2.1.1
fasteners==0.18
packaging==23.1
python-dotenv==1.0.0
setuptools-scm==7.1.0
typer==0.9.0
typing_extensions==4.7.1
